// src/interfaces/aclwallets-data.ts
export interface AclwalletData {
  id: number;
  uuid: string;
  nacionality: string;
  name: string;
  document: string;
  phone: string;
  zipCode: string;
  address: string;
  number: string;
  country: string;
  wallet: string;
  createdAt: string;
  updateAt: string;
  email: string;
  authenticationId: number;
  networkId: number;
  active: boolean;
  externalTransfer: boolean;
  accountId: number;
  btcWallet: any;
  liquidWallet: any;
  freeRefund: boolean;
}